dimensys-explorer-main/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
httpx>=0.27.0
python-dotenv>=1.0.0
transformers>=4.38.0
torch>=2.1.0
sentence-transformers>=2.6.0
numpy>=1.26.0
